- Fix type conversion warning in file_wad.cpp.
SVN r3992 (trunk)

+++ src/resourcefiles/file_wad.cpp
@@ -220,7 +220,7 @@ public:
 		while(AvailOut && Stream.State != STREAM_FINAL);
 
 		assert(AvailOut == 0);
-		return Out - (BYTE*)buffer;
+		return (long)(Out - (BYTE*)buffer);
 	}
 };
